Remove unnecessary files and configurations
