+++ demos/candid-react/src/declarations/simple/b3_system.did
