+++ demos/candid-react/src/declarations/simple/index.d.ts
