+++ demos/candid-react/src/declarations/simple/index.js
