--- demos/candid-types/.eslintignore
@@ -1,6 +1,0 @@
-# don't ever lint node_modules
-node_modules
-# don't lint build output (make sure it's set to your correct build folder name)
-build
-# don't lint nyc coverage output
-coverage
