--- demos/candid-types/.eslintrc
@@ -1,19 +1,0 @@
-{
-  "root": true,
-  "parser": "@typescript-eslint/parser",
-  "plugins": ["@typescript-eslint", "prettier", "jest"],
-  "extends": [
-    "eslint:recommended",
-    "plugin:@typescript-eslint/recommended",
-    "prettier"
-  ],
-  "rules": {
-    "no-console": 1,
-    "prettier/prettier": 2
-  },
-  "env": {
-    "browser": true,
-    "node": true,
-    "jest/globals": true
-  }
-}
