--- demos/candid-types/.prettierrc
@@ -1,5 +1,0 @@
-{
-  "trailingComma": "all",
-  "singleQuote": true,
-  "printWidth": 80
-}
