--- demos/candid-types/jest.config.js
@@ -1,9 +1,0 @@
-module.exports = {
-  transform: {
-    '^.+\\.ts?$': 'ts-jest',
-  },
-  testEnvironment: 'node',
-  testRegex: './src/.*\\.(test|spec)?\\.(ts|ts)$',
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
-  roots: ['<rootDir>/src'],
-};
