--- demos/candid-types/launch.json
@@ -1,17 +1,0 @@
-{
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "type": "node",
-      "request": "launch",
-      "name": "Debug TypeScript in Node.js",
-      "preLaunchTask": "npm: build",
-      "program": "${workspaceFolder}/src/index.ts",
-      "protocol": "inspector",
-      "outFiles": ["${workspaceFolder}/build/**/*.js"],
-      "sourceMaps": true,
-      "smartStep": true,
-      "internalConsoleOptions": "openOnSessionStart"
-    }
-  ]
-}
