--- demos/candid-types/tasks.json
@@ -1,12 +1,0 @@
-{
-  // See https://go.microsoft.com/fwlink/?LinkId=733558
-  // for the documentation about the tasks.json format
-  "version": "2.0.0",
-  "tasks": [
-    {
-      "type": "typescript",
-      "tsconfig": "simple-typescript-starter/tsconfig.json",
-      "problemMatcher": ["$tsc"]
-    }
-  ]
-}
