--- demos/candid-types/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es5",
-    "module": "commonjs",
-    "lib": ["es6"],
-    "allowJs": true,
-    "outDir": "build",
-    "rootDir": "src",
-    "strict": true,
-    "baseUrl": ".",
-    "esModuleInterop": true,
-    "types": ["node", "jest"],
-    "skipLibCheck": true,
-    "sourceMap": true
-  },
-  "include": ["src/**/*"],
-  "exclude": ["src/**/*.spec.ts", "build/**/*"]
-}
